requirements.txt
neovim
django==4.0
psycopg2==2.9.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0 
django-filter==22.1

django-extensions==3.2.1

 django-environ==0.9.0 
